# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: NinerLog API
  version: 1.0.0
  description: |
    RESTful API for EASA/FAA compliant pilot logbook management.
    
    ## Features
    - Multi-license support (PPL, SPL, CPL, ATPL, IR)
    - EASA and FAA regulatory compliance
    - Flight logging with detailed time tracking
    - Currency calculations
    - Statistics and hour totals per license
    
    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
  contact:
    name: Frederic Jaeckel
    email: frederic@ninerlog.app
  license:
    name: TBD
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://app.ninerlog.com/api/v1
    description: Production server
...
paths:
...
  /auth/2fa/setup:
    post:
      tags:
        - Authentication
      summary: Start 2FA setup
      description: Generate a TOTP secret and return the provisioning URI for authenticator apps. Does not enable 2FA until verified.
      operationId: setup2FA
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 2FA setup data with secret and QR URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorSetup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 2FA is already enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/LocalAuthDisabled'
  /auth/2fa/verify:
    post:
      tags:
        - Authentication
      summary: Verify and enable 2FA
      description: Verify a TOTP code to confirm setup and enable 2FA. Returns backup recovery codes.
      operationId: verify2FA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  description: 6-digit TOTP code from authenticator app
      responses:
        '200':
          description: 2FA enabled with recovery codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorEnabled'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/LocalAuthDisabled'
  /auth/2fa/disable:
    post:
      tags:
        - Authentication
      summary: Disable 2FA
      description: Disable two-factor authentication. Requires current password for confirmation.
      operationId: disable2FA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
      responses:
        '204':
          description: 2FA disabled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/LocalAuthDisabled'
  /auth/2fa/login:
    post:
      tags:
        - Authentication
      summary: Complete login with 2FA code
      description: After initial login returns requiresTwoFactor, submit the TOTP code or recovery code here.
      operationId: login2FA
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - twoFactorToken
                - code
              properties:
                twoFactorToken:
                  type: string
                  description: Temporary token from login response
                code:
                  type: string
                  description: 6-digit TOTP code or recovery code
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/LocalAuthDisabled'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
...
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: pilot@example.com
        name:
          type: string
          example: John Doe
        createdAt:
          type: string
          format: date-time
          example: '2026-01-15T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-20T14:45:00Z'
        twoFactorEnabled:
          type: boolean
          example: false
          description: Whether 2FA is enabled for this account
        emailVerified:
          type: boolean
          example: true
          description: Whether the user has confirmed their email address.
        isAdmin:
          type: boolean
          example: false
          description: Whether this user is the platform admin (computed from ADMIN_EMAIL env var, not stored in DB)
        timeDisplayFormat:
          type: string
          enum: [hm, decimal]
          default: hm
          example: hm
          description: User's preferred time display format. "hm" for hours and minutes (1h 30m), "decimal" for decimal hours (1.5h).
        dateFormat:
          type: string
          enum: [DD.MM.YYYY, MM/DD/YYYY, YYYY-MM-DD]
          default: DD.MM.YYYY
          example: DD.MM.YYYY
          description: User's preferred date display format.
        decimalSeparator:
          type: string
          enum: [comma, dot]
          default: comma
          example: comma
          description: User's preferred decimal separator for number display. "comma" for European style (1,5), "dot" for US/UK style (1.5).
        preferredLocale:
          type: string
          enum: [en, de]
          default: en
          example: en
          description: User's preferred language for the interface.
        recencyPerModel:
          type: boolean
          default: true
          example: true
          description: >-
            Show informational 90-day landing recency (styled after EASA
            FCL.060(b)) per aircraft model/type.
        recencyPerRegistration:
          type: boolean
          default: false
          example: false
          description: >-
            Show informational 90-day landing recency per individual aircraft
            registration. Off by default — FCL.060(b) recency is defined per
            type/class, so per-registration tracking is purely a familiarity
            aid.
        flightListColumnMode:
          type: string
          enum: [auto, custom]
          default: auto
          example: auto
          description: >-
            How the optional columns of the flights list are chosen. "auto"
            lets the client pick them from the flights on the page, so an
            unused column (IFR time for a VFR-only pilot) never takes up width.
            "custom" means flightListColumns is the user's own choice.
        flightListColumns:
          type: array
          items:
            $ref: '#/components/schemas/FlightListColumn'
          default: []
          example: [picTime, nightTime, landings]
          description: >-
            The optional flights-list columns the user picked. Only meaningful
            when flightListColumnMode is "custom", where an empty array means
            "none of the optional columns". Stored deduplicated and in the
            canonical display order; unknown values are dropped.
    FlightListColumn:
      type: string
      description: >-
        An optional column of the flights list. Date, route, aircraft and total
        time are the identity of a logbook row and are always shown, so they
        are not part of this enum. The order below is the display order, and
        for the time columns also the priority order in which they survive as
        the list gets narrower.
      enum:
        - offOnBlock
        - picTime
        - nightTime
        - dualTime
        - ifrTime
        - crossCountryTime
        - sicTime
        - dualGivenTime
        - multiPilotTime
        - soloTime
        - simulatedFlightTime
        - function
        - landings
        - remarks
      example: picTime
    TwoFactorSetup:
      type: object
      required:
        - secret
        - qrUri
      properties:
        secret:
          type: string
          description: Base32-encoded TOTP secret for manual entry
          example: JBSWY3DPEHPK3PXP
        qrUri:
          type: string
          description: otpauth:// URI for QR code generation
          example: 'otpauth://totp/NinerLog:pilot@example.com?secret=JBSWY3DPEHPK3PXP&issuer=NinerLog'
...
    TwoFactorEnabled:
      type: object
      required:
        - recoveryCodes
      properties:
        recoveryCodes:
          type: array
          items:
            type: string
          description: One-time recovery codes to use if authenticator is unavailable
          example: ['abc12-def34', 'ghi56-jkl78', 'mno90-pqr12']
...
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          description: JWT access token (15 minutes validity)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token (7 days validity)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Access token expiry in seconds
          example: 900
        user:
          $ref: '#/components/schemas/User'
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Validation failed
          description: Error message
        code:
          type: string
          example: email_not_verified
          description: |
            Optional machine-readable error code, used by clients to distinguish
            between conditions that share an HTTP status (e.g. `email_not_verified`
            vs `account_disabled` for 403 on /auth/login).
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: totalTime
              message:
                type: string
                example: Must be greater than 0
          description: Detailed validation errors
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            details:
              - field: totalTime
                message: Must be greater than 0
              - field: nightTime
                message: Cannot exceed total time
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
...
    LocalAuthDisabled:
      description: >-
        The server runs in OIDC mode, where the identity provider owns all
        accounts and credentials. Every local-credential endpoint — passwords,
        registration, email verification, TOTP and passkeys — is switched off.
        Clients should call GET /auth/providers and render the OIDC sign-in
        path instead.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Local authentication is disabled on this server (OIDC mode)